--- frb/data/Galaxies/180301/FRB180301_host.json
@@ -1,92 +1,0 @@
-{
-    "FRB": "FRB180301",
-    "cosmo": "Planck15",
-    "dec": 4.670476944444445,
-    "dec_FRB": 4.671083333333334,
-    "derived": {
-        "AV_nebular": -0.7320378963820574,
-        "EBV_photom": 0.22627995267562478,
-        "EBV_photom_err": 0.09007260242919658,
-        "EBV_spec": 0.9698040690863088,
-        "Lnu_r": 5.153892186015295e+21,
-        "Lnu_r_err": 3.090556468877778e+20,
-        "M_r": -20.18033832331934,
-        "M_r_err": 0.06510669858820534,
-        "Mstar": 2301186534.4534163,
-        "Mstar_err": 648156843.0347536,
-        "Mstar_spec": 92965026674.1868,
-        "SFR_nebular": 1.9261730063547269,
-        "SFR_photom": 3.1149155410630684,
-        "SFR_photom_err": 0.9971522203744531,
-        "Z_photom": 0.004439210459498475,
-        "Z_photom_err": 0.0074121206939024116,
-        "Z_spec": 0.2113762757801206,
-        "age_mass": 607.2427044007321,
-        "f_AGN": 0.1485917800193184,
-        "f_AGN_err": 0.06875660111491497,
-        "u-r": 0.8980443529309934,
-        "u-r_err": 0.11052647794409519
-    },
-    "neb_lines": {
-        "Halpha": 1.6966844254164578e-15,
-        "Halpha_err": 1.7532910003576657e-17,
-        "Hbeta": 7.77048982264961e-16,
-        "Hbeta_err": 6.072268912520871e-17,
-        "Hgamma": 2.722240094399401e-16,
-        "Hgamma_err": 7.022315864308002e-17,
-        "[NII] 6584": 4.3214761416231567e-16,
-        "[NII] 6584_err": 3.239333918843798e-17,
-        "[OIII] 5007": 1.2747689345291548e-15,
-        "[OIII] 5007_err": 6.461255110876571e-17
-    },
-    "offsets": {
-        "ang_avg": 2.2866372154619463,
-        "ang_avg_err": 0.6050873980879884,
-        "ang_best": 2.201790863836011,
-        "ang_best_err": 0.6110070889188792,
-        "physical": 10.797974349555423,
-        "physical_err": 2.9964875329022065
-    },
-    "photom": {
-        "EBV": 0.46,
-        "MMIRS_H": 20.402285737626148,
-        "MMIRS_H_err": 0.0827,
-        "MMIRS_H_flux": 0.025066038318689157,
-        "MMIRS_H_flux_err": 0.001983863881771235,
-        "MMIRS_J": 20.56447143683972,
-        "MMIRS_J_err": 0.066,
-        "MMIRS_J_flux": 0.02158795079714274,
-        "MMIRS_J_flux_err": 0.0013530003225751563,
-        "MMIRS_K": 20.511774213420054,
-        "MMIRS_K_err": 0.076,
-        "MMIRS_K_flux": 0.02266158640084691,
-        "MMIRS_K_flux_err": 0.0016431160674875885,
-        "NOT_g": 21.40293355446113,
-        "NOT_g_err": 0.09,
-        "NOT_g_flux": 0.009973017302448893,
-        "NOT_g_flux_err": 0.0008619242397079302,
-        "NOT_i": 20.992619681418205,
-        "NOT_i_err": 0.06,
-        "NOT_i_flux": 0.01455298634089752,
-        "NOT_i_flux_err": 0.0008268644355101907,
-        "NOT_r": 21.041589219904807,
-        "NOT_r_err": 0.06,
-        "NOT_r_flux": 0.013911190770430952,
-        "NOT_r_flux_err": 0.0007903992097719213,
-        "NOT_u": 21.457713823811197,
-        "NOT_u_err": 0.3,
-        "NOT_u_flux": 0.009482316886426596,
-        "NOT_u_flux_err": 0.0030178112462324724,
-        "NOT_z": 20.585880944956735,
-        "NOT_z_err": 0.11,
-        "NOT_z_flux": 0.02116643008078041,
-        "NOT_z_flux_err": 0.0022568448685115075
-    },
-    "ra": 93.22682999999999,
-    "ra_FRB": 93.22675,
-    "redshift": {
-        "z": 0.3304,
-        "z_FRB": 0.3304,
-        "z_spec": 0.3304
-    }
-}
