--- frb/data/Galaxies/180916/FRB180916_host.json
@@ -1,112 +1,0 @@
-{
-    "FRB": "FRB180916",
-    "cosmo": "Planck15",
-    "dec": 65.71474666666667,
-    "dec_FRB": 65.71675422222222,
-    "derived": {
-        "EBV_photom": 0.12000000000047942,
-        "EBV_photom_err": 2.4965527214465155e-07,
-        "Lnu_r": 2.6549127112243004e+21,
-        "Lnu_r_err": 1.3274563556121503e+20,
-        "M_r": -19.460125617068393,
-        "M_r_err": 0.05428681023790647,
-        "Mstar": 2150807243.957849,
-        "Mstar_err": 330647885.4821547,
-        "SFR_nebular": 0.05610366390201721,
-        "SFR_photom": 0.0025570380687979905,
-        "SFR_photom_err": 0.00971244278020802,
-        "Z_photom": 0.00039352129597312934,
-        "Z_photom_err": 4.379418888252133e-05,
-        "age_mass": 154.85148781723885,
-        "f_AGN": 0.004327549308433153,
-        "f_AGN_err": 0.014600948419969667,
-        "u-r": 1.5283007538257023,
-        "u-r_err": 0.06187590927411147
-    },
-    "morphology": {
-        "PA": 63.76,
-        "PA_err": 0.08,
-        "b/a": 0.7,
-        "b/a_err": 0.0,
-        "dec": 65.71482219472851,
-        "dec_err": 0.005135184298127717,
-        "n": 0.64,
-        "n_err": 0.0,
-        "ra": 29.501123751074392,
-        "ra_err": 0.007702776447191576,
-        "reff_ang": 8.648934154121608,
-        "reff_ang_err": 0.017973145043447013,
-        "reff_kpc": 6.002340348201615,
-        "reff_kpc_err": 0.012473321192641032
-    },
-    "neb_lines": {
-        "Halpha": 4.026875885927211e-15,
-        "Halpha_err": 2.45167481639404e-17,
-        "[NII] 6584": 1.5243034753395692e-15,
-        "[NII] 6584_err": 2.4291688850032977e-17,
-        "[OIII] 4959": 5.908030232542076e-16,
-        "[OIII] 4959_err": 6.218979192149554e-17,
-        "[OIII] 5007": 7.161580805676158e-15,
-        "[OIII] 5007_err": 6.0181351308203e-17
-    },
-    "offsets": {
-        "ang_avg": 7.760248229369667,
-        "ang_avg_err": 0.0225899311120113,
-        "ang_best": 7.760227588300796,
-        "ang_best_err": 0.02258994054217806,
-        "physical": 5.385580041939379,
-        "physical_err": 0.015677366616922952
-    },
-    "photom": {
-        "EBV": 0.8702,
-        "SDSS_g": 17.075671999999997,
-        "SDSS_g_err": 0.08271651,
-        "SDSS_g_flux": 0.536699504454394,
-        "SDSS_g_flux_err": 0.042486152379357434,
-        "SDSS_i": 15.933784000000001,
-        "SDSS_i_err": 0.02263959,
-        "SDSS_i_flux": 1.5363366617691698,
-        "SDSS_i_flux_err": 0.03237176789245194,
-        "SDSS_r": 16.170141,
-        "SDSS_r_err": 0.03178847,
-        "SDSS_r_flux": 1.2357869203292875,
-        "SDSS_r_flux_err": 0.036716570189002574,
-        "SDSS_u": 20.314096,
-        "SDSS_u_err": 1.781266,
-        "SDSS_u_flux": 0.02718701922420491,
-        "SDSS_u_flux_err": 0.11305171857589894,
-        "SDSS_z": 15.851336,
-        "SDSS_z_err": 0.06344854,
-        "SDSS_z_flux": 1.6575460174286705,
-        "SDSS_z_flux_err": 0.09975034598134898,
-        "WISE_W1": 14.367767338885313,
-        "WISE_W1_err": 0.03,
-        "WISE_W1_flux": 0.5541276773910984,
-        "WISE_W1_flux_err": 0.01552460620991937,
-        "WISE_W2": 14.41414748838363,
-        "WISE_W2_err": 0.051,
-        "WISE_W2_flux": 0.2946668633010211,
-        "WISE_W2_flux_err": 0.014171541154414553,
-        "WISE_W3": 10.558888586015737,
-        "WISE_W3_err": 0.079,
-        "WISE_W3_flux": 1.8929863309359283,
-        "WISE_W3_flux_err": 0.14287164592369073,
-        "WISE_W4": 9.078791625036564,
-        "WISE_W4_err": 0.522,
-        "WISE_W4_flux": 1.953644084895555,
-        "WISE_W4_flux_err": 1.206052980162046
-    },
-    "positional_error": {
-        "dec_astrometric": 0.0232,
-        "dec_source": 0.00108,
-        "ra_astrometric": 0.0167,
-        "ra_source": 0.0014
-    },
-    "ra": 29.501216666666668,
-    "ra_FRB": 29.503125708333332,
-    "redshift": {
-        "z": 0.0337,
-        "z_FRB": 0.0337,
-        "z_spec": 0.0337
-    }
-}
