+++ frb/data/Galaxies/20180301/FRB20180301_host.json
@@ -5,10 +5,43 @@
     "dec_FRB": 4.671083333333334,
     "derived": {
         "AV_nebular": -0.7320378963820574,
+        "EBV_photom": 0.22627995267562478,
+        "EBV_photom_err": 0.09007260242919658,
         "EBV_spec": 0.9698040690863088,
+        "Lnu_r": 5.153892186015295e+21,
+        "Lnu_r_err": 3.090556468877778e+20,
+        "M_r": -20.18033832331934,
+        "M_r_err": 0.06510669858820534,
+        "Mstar": 2301186534.4534163,
+        "Mstar_err": 648156843.0347536,
         "Mstar_spec": 92965026674.1868,
         "SFR_nebular": 1.9301083213653862,
-        "Z_spec": 0.2113762757801206
+        "SFR_photom": 3.1149155410630684,
+        "SFR_photom_err": 0.9971522203744531,
+        "Z_photom": 0.004439210459498475,
+        "Z_photom_err": 0.0074121206939024116,
+        "Z_spec": 0.2113762757801206,
+        "age_mass": 607.2427044007321,
+        "f_AGN": 0.1485917800193184,
+        "f_AGN_err": 0.06875660111491497,
+        "u-r": 0.8980443529309934,
+        "u-r_err": 0.11052647794409519
+    },
+    "morphology": {
+        "PA": -9.69,
+        "PA_err": 1.01,
+        "b/a": 0.33,
+        "b/a_err": 0.01,
+        "dec": 4.67036313419657,
+        "dec_err": 0.01919986693935617,
+        "n": 0.71,
+        "n_err": 0.07,
+        "ra": 93.22685063035637,
+        "ra_err": 0.009599933469678084,
+        "reff_ang": 1.1903917502400827,
+        "reff_ang_err": 0.03039978932064727,
+        "reff_kpc": 5.8429752222267535,
+        "reff_kpc_err": 0.14921576508374773
     },
     "neb_lines": {
         "Halpha": 1.6966844254164578e-15,
