+++ frb/data/Galaxies/20180916/FRB20180916_host.json
@@ -3,6 +3,22 @@
     "cosmo": "Planck18",
     "dec": 65.71474666666667,
     "dec_FRB": 65.71675422222222,
+    "morphology": {
+        "PA": 63.76,
+        "PA_err": 0.08,
+        "b/a": 0.7,
+        "b/a_err": 0.0,
+        "dec": 65.71482219472851,
+        "dec_err": 0.005135184298127717,
+        "n": 0.64,
+        "n_err": 0.0,
+        "ra": 29.501123751074392,
+        "ra_err": 0.007702776447191576,
+        "reff_ang": 8.648934154121608,
+        "reff_ang_err": 0.017973145043447013,
+        "reff_kpc": 6.009109277517017,
+        "reff_kpc_err": 0.012487387544194484
+    },
     "offsets": {
         "ang_avg": 7.760248229369667,
         "ang_avg_err": 0.0225899311120113,
