+++ frb/data/Galaxies/20191228/FRB20191228_host.json
@@ -27,6 +27,22 @@
         "u-r": 2.1324868391184695,
         "u-r_err": 0.7422042727886725
     },
+    "morphology": {
+        "PA": -66.52,
+        "PA_err": 3.75,
+        "b/a": 0.6,
+        "b/a_err": 0.03,
+        "dec": -29.59420391092428,
+        "dec_err": 0.007558596000000001,
+        "n": 0.29,
+        "n_err": 0.17,
+        "ra": 344.43085441057696,
+        "ra_err": 0.010078128000000002,
+        "reff_ang": 0.4509962280000001,
+        "reff_ang_err": 0.015117192000000002,
+        "reff_kpc": 1.7840279446047402,
+        "reff_kpc_err": 0.05979981937222592
+    },
     "neb_lines": {
         "Halpha": 3.004182313912709e-17,
         "Halpha_err": 1.776185487993526e-18,
